Fix test_policy & notebooks
- Correct test_policy to use updated methods of SimulationModel class
- Add testing logs on new kql_logs folder
- Fix sim validation notebook to use correct SimulatorModel methods
- Fix order of CSVs on analysis JNB to ensure correct pairwise plot per-concept
- Isolate log functionality into unique class for easy embed on new sims
